+++ src/masking.rs
@@ -1,4 +1,4 @@
-use std::iter::repeat;
+use std::iter::repeat_n;
 
 /// Default replacement value for `#[deboog(mask = "hidden")]`
 pub const HIDE_STR: &str = "***";
@@ -35,7 +35,7 @@ pub fn mask_all(value: &str) -> String {
 /// Leaves 6 characters in the beginning and 4 characters at the end of the string.
 pub fn mask_pan(value: &str) -> String {
     let sz = value.len();
-    let limit = if sz > 4 { sz - 4 } else { 0 };
+    let limit = sz.saturating_sub(4);
     value
         .chars()
         .enumerate()
@@ -46,9 +46,8 @@ pub fn mask_pan(value: &str) -> String {
 /// Replaces all but last four characters with a single `*` symbol
 pub fn mask_pan_suffix(value: &str) -> String {
     let sz = value.len();
-    let limit = if sz > 4 { sz - 4 } else { 0 };
-    repeat('*')
-        .take(if sz > 4 { 1 } else { 0 })
+    let limit = sz.saturating_sub(4);
+    repeat_n('*', if sz > 4 { 1 } else { 0 })
         .chain(value[limit..].chars())
         .collect()
 }