+++ tests/basic.rs
@@ -9,20 +9,23 @@ fn unit_struct() {
 
 #[test]
 fn tuple_struct() {
+    #[allow(dead_code)]
     #[derive(Deboog)]
     struct Tuple(i32, &'static str);
     let our = Tuple(123, "test");
-    assert_eq!(format!("{:?}", our), r#"Tuple(123, "test")"#);
+    assert_eq!(format!("{our:?}"), r#"Tuple(123, "test")"#);
 }
 
 #[test]
 fn nested_tuple_struct() {
+    #[allow(dead_code)]
     #[derive(Deboog)]
     struct Inner(i32);
+    #[allow(dead_code)]
     #[derive(Deboog)]
     struct Outer(Inner);
     let our = Outer(Inner(123));
-    assert_eq!(format!("{:?}", our), r#"Outer(Inner(123))"#);
+    assert_eq!(format!("{our:?}"), r#"Outer(Inner(123))"#);
 }
 
 #[test]
@@ -34,7 +37,7 @@ fn normal_struct() {
         b: &'static str,
     }
     let our = Struct { a: 123, b: "test" };
-    assert_eq!(format!("{:?}", our), r#"Struct { a: 123, b: "test" }"#);
+    assert_eq!(format!("{our:?}"), r#"Struct { a: 123, b: "test" }"#);
 }
 
 #[test]
@@ -53,7 +56,7 @@ fn nested_normal_struct() {
     let inner = Inner { a: 123, b: "test" };
     let our = Outer { inner };
     assert_eq!(
-        format!("{:?}", our),
+        format!("{our:?}"),
         r#"Outer { inner: Inner { a: 123, b: "test" } }"#
     );
 }