+++ CHANGELOG.md
@@ -2,10 +2,20 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2025-10-07
+
+### Added
+
+- `chrono` support via a feature flag.
+
+### Fixed
+
+- Fixed broken derive macro dependency.
 
 ## [0.2.0] - 2023-07-25
 
@@ -36,7 +46,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial code push.
 
-[Unreleased]: https://github.com/unikmhz/deboog/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/unikmhz/deboog/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/unikmhz/deboog/releases/tag/v0.3.0
 [0.2.0]: https://github.com/unikmhz/deboog/releases/tag/v0.2.0
 [0.1.1]: https://github.com/unikmhz/deboog/releases/tag/v0.1.1
 [0.1.0]: https://github.com/unikmhz/deboog/releases/tag/v0.1.0